feat: add filtering and log priority functionality to Android native logs (#4)
* feat: Add filtering functionality to Android logs

This commit adds filtering functionality to the `readOutputLogs` method in the Android native logs module. The new version of the method now accepts an optional array of tags to ignore when reading logs from the file.

The commit also updates the `redirectLogs` method to accept a shortened version of the log level string instead, allowing also to add filters based on the log level priority.

Both changes are specific to Android devices and improve the functionality of the native logs module.

* chore: add method overloading for logging functionality

This commit adds method overloading for logging functionality. The overloading was introduced in two new methods: `setUpRedirectLogs` and `readOutputLogs`.

These methods provide the same functionality as their counterparts but with additional parameters that allow more control over the logging process.

This overloading was added to ensure backward compatibility with iOS apps and prevent breaking changes.

* chore: add method overloading and platform checks

* fix: add filtering tags as optional parameter for getLogs

* chore: add missing types

* chore: add optional parameters type description and simplify conditions.